requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.28
pydantic==2.6.4
pydantic-settings==2.2.1
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
email-validator==2.1.1
python-multipart==0.0.9
bcrypt==3.2.2
httpx==0.27.2